Updated Trezor plugin
Adapted to current Trezor library
Cashaddr support
Trezor T support
Matrix recovery support
Show multisig address on Trezor
